Move GenericRecord to AnimeManga folder.
