+++ Atarashii/src/net/somethingdreadful/MAL/api/response/AnimeManga/Manga.java
@@ -5,7 +5,6 @@ import android.database.Cursor;
 import com.google.gson.Gson;
 import com.google.gson.annotations.SerializedName;
 
-import net.somethingdreadful.MAL.api.response.GenericRecord;
 import net.somethingdreadful.MAL.api.response.RecordStub;
 import net.somethingdreadful.MAL.sql.MALSqlHelper;
