+++ Atarashii/src/net/somethingdreadful/MAL/tasks/WriteDetailTask.java
@@ -15,7 +15,7 @@ import net.somethingdreadful.MAL.api.MALApi;
 import net.somethingdreadful.MAL.api.MALApi.ListType;
 import net.somethingdreadful.MAL.api.response.AnimeManga.Anime;
 import net.somethingdreadful.MAL.api.response.AnimeManga.Manga;
-import net.somethingdreadful.MAL.api.response.GenericRecord;
+import net.somethingdreadful.MAL.api.response.AnimeManga.GenericRecord;
 
 import retrofit.RetrofitError;
